--- a/pkg/stanza/operator/parser/json/config.go
+++ b/pkg/stanza/operator/parser/json/config.go
@@ -4,7 +4,6 @@
 package json // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/json"
 
 import (
-	jsoniter "github.com/json-iterator/go"
 	"go.opentelemetry.io/collector/component"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
@@ -43,6 +42,5 @@ func (c Config) Build(set component.TelemetrySettings) (operator.Operator, error
 
 	return &Parser{
 		ParserOperator: parserOperator,
-		json:           jsoniter.ConfigFastest,
 	}, nil
 }