--- a/pkg/stanza/operator/parser/json/parser.go
+++ b/pkg/stanza/operator/parser/json/parser.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 
 	"github.com/goccy/go-json"
+
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )