**Description:** 
Switch libraries to reduce the footprint of the JSON parser.

**Link to tracking Issue:**
Fixes #33784 

**Testing:**

Before:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/json
BenchmarkProcess-10        49047             23951 ns/op           26277 B/op        177 allocs/op
BenchmarkProcess-10        50248             24002 ns/op           26275 B/op        177 allocs/op
BenchmarkProcess-10        50258             24517 ns/op           26276 B/op        177 allocs/op
BenchmarkProcess-10        50505             24731 ns/op           26276 B/op        177 allocs/op
BenchmarkProcess-10        45708             24730 ns/op           26276 B/op        177 allocs/op
BenchmarkProcess-10        50022             25021 ns/op           26277 B/op        177 allocs/op
BenchmarkProcess-10        47204             24794 ns/op           26275 B/op        177 allocs/op
BenchmarkProcess-10        47742             25335 ns/op           26274 B/op        177 allocs/op
BenchmarkProcess-10        46252             25205 ns/op           26276 B/op        177 allocs/op
BenchmarkProcess-10        47916             24379 ns/op           26277 B/op        177 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/json       15.559s
```

After:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/json
BenchmarkProcess-10        95613             12168 ns/op           14702 B/op        115 allocs/op
BenchmarkProcess-10        97904             12343 ns/op           14704 B/op        115 allocs/op
BenchmarkProcess-10        99006             12187 ns/op           14702 B/op        115 allocs/op
BenchmarkProcess-10        96964             12310 ns/op           14703 B/op        115 allocs/op
BenchmarkProcess-10        98661             12285 ns/op           14703 B/op        115 allocs/op
BenchmarkProcess-10        96896             12356 ns/op           14703 B/op        115 allocs/op
BenchmarkProcess-10        94118             12367 ns/op           14703 B/op        115 allocs/op
BenchmarkProcess-10        96123             12349 ns/op           14702 B/op        115 allocs/op
BenchmarkProcess-10        96948             12305 ns/op           14702 B/op        115 allocs/op
BenchmarkProcess-10        96626             12225 ns/op           14702 B/op        115 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/json       13.619s
```

Benchstat:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/json
           │ before.txt  │              after.txt              │
           │   sec/op    │   sec/op     vs base                │
Process-10   24.73µ ± 3%   12.31µ ± 1%  -50.23% (p=0.000 n=10)

           │  before.txt  │              after.txt               │
           │     B/op     │     B/op      vs base                │
Process-10   25.66Ki ± 0%   14.36Ki ± 0%  -44.05% (p=0.000 n=10)

           │ before.txt │             after.txt              │
           │ allocs/op  │ allocs/op   vs base                │
Process-10   177.0 ± 0%   115.0 ± 0%  -35.03% (p=0.000 n=10)
```