@andrzej-stencel 2024-06-27
Oh, it looks like the https://github.com/json-iterator/go that is currently used has been unmaintained for about 2 years now?

@atoulme are you aware of other major open source projects using the proposed library https://github.com/goccy/go-json?