@atoulme 2024-06-27
> @atoulme are you aware of other major open source projects using the proposed library https://github.com/goccy/go-json?
I took the first google result. I applied it to our code, and I saw it is used by quite a few other components. 

```
 git grep -l go-json | xargs dirname | sort | uniq
cmd/otelcontribcol
connector/datadogconnector
exporter/datadogexporter
exporter/datadogexporter/integrationtest
exporter/otelarrowexporter
internal/sqlquery
processor/k8sattributesprocessor
receiver/googlecloudspannerreceiver
receiver/influxdbreceiver
receiver/otelarrowreceiver
receiver/snowflakereceiver
receiver/sqlqueryreceiver
receiver/sqlserverreceiver
```

Following a few of those paths, it seems to be used by datadog libraries and the Go arrow client.

Given that json-iterator/go is not seeing any updates for 2 years, jumping to another lib with minimal change to the functional set seems like the right idea.